Removed extra space in search bar placeholder
I removed an extra space, showing in the space bar placeholder.
It was showing `Enter name of song  or YouTube URL` with two space characters between `song` and `or`.

Because it is a simple typo fix, I didn't discuss it first via issue or email. I hope that's alright.
Thanks for your work!