Fix the white screen issue
The white screen on second load was happenning due to nginx and not our
code. nginx (since it redirects all 404's to our index was doing just
that.). This means, even when a chunk that the service worker has cached
wouldn't be available, it'd be returned the `index.html` with a 200 OK
and so our service worker thinks that's the correct file and accordingly
caches it.

In the second load, it tries to load this cached files and fails in
doing so since the JS and CSS files are actually the `index.html` file.