Once all the covers are fetched, pass them to the playlist component
Once all the covers are individually fetched then they are emitted to
the parent. This parent makes an array of all the emitted covers and
sends it to the parent which is the playlist component.

This component will pass the cover to the head which will build a
cover image for the playlist